dpdk: fix CVE-2024-11614
Issue: LIN1025-436

An out-of-bounds read vulnerability was found in DPDK's Vhost
library checksum offload feature. This issue enables an
untrusted or compromised guest to crash the hypervisor's vSwitch
by forging Virtio descriptors to cause out-of-bounds reads. This
flaw allows an attacker with a malicious VM using a virtio driver
to cause the vhost-user side to crash by sending a packet with a
Tx checksum offload request and an invalid csum_start offset.

Since the 'walnascar' branch doesn't exist in meta-dpdk, I'm sending the RR

Reference:
https://security-tracker.debian.org/tracker/CVE-2024-11614

Upstream-patch:
https://git.dpdk.org/dpdk/commit/?id=4dc4e33ffa108e945fc8a1e2bbc7819791faa61e

(LOCAL REV: NOT UPSTREAM) -- this fix is wrlinux specific

Signed-off-by: Divya Chellam <divya.chellam@windriver.com>